Added consistent spacing
Hello,

I have made a modification to the GitHub project. I added spaces to make the text consistent throughout the message.yml file. Previously, there was no consistency in the formatting of the text, making it difficult to read and understand.

Therefore, I have corrected this by adding spaces to align the text in a consistent and uniform manner. This modification will allow users to better understand messages and have a more pleasant user experience.

I invite you to review the modification and provide me with your feedback if necessary.

Best regards,